refactor(descriptor): Drop many properties from the top-level struct (#98)
- Remove most properties from `DataPackage` (descriptor) and `DataResource` structs

`DataPackage` was initially bootstrapped from [Frictionless Data Package descriptors](https://specs.frictionlessdata.io/data-package/). Frictionless descriptors were designed to be
- fairly self-contained, making few references to anything external except the data source itself; and
- fairly concrete descriptions of data, down to physical details like size in bytes, hash of the files containing the data, and serialization details of the data in each field

A lot of that doesn't carry over to the system architecture of DPM Cloud, where the descriptor document is a relatively abstract description of data in a remote location and queries are resolved by agent.dpm.sh. Furthermore, the serialization format of data is only an internal contract between the agent's query API and the package instances that consume it; the software using a package only ever receives deserialized data.

This PR only touches the topmost couple structs. Similar simplification on the table schema / field def structs will come in a subsequent PR.