test: Run integration tests on oldest supported targets (#105)
- Avoid using the `|` operator in Python codegen, as that is unsupported
until 3.10 (see [PEP 604](https://peps.python.org/pep-0604/))
- Run integration tests on the oldest supported major release lines of
targets, to detect bugs such as this

There's an argument to be made that integration tests, or at least the
tests of the built package instances, should be run against the same
matrix of target versions as the static tests are run under. This PR
does not pursue that.

## Test plan

The two commits in this PR, their check runs demonstrate the bug being
detected and then being fixed.