feat(pat-3891): use get package version query to build data packages (#113)
Replaces the descriptor file input on `dpm build-package` with a package
identifier (package_name@version), used to query dpm cloud for the
dataset info which builds the package. This makes the descriptor stored
in dpm cloud the exclusive source of truth, mostly eliminating the "how
do we sync the descriptor from place x system to place y complexity."

**Test Plan:**
1. Created a new account with dpm login, then ran the command: `dpm
build-package -p "Snowflake Demo Package@0.1.0" -y python`, validating
that generated artifacts were created and corresponded to the data in
the `dataset` column of postgres.
2. Created a new source, described a table on said source, then used
`dpm publish` to create a version 0.1.0 of a `test-package` data
package. Then ran `dpm build-package -p test-package@0.1.0 -y nodejs`,
validating that generated artifacts were created and corresponded to the
data in the `dataset` column of postgres.

---------

Co-authored-by: Spencer Wilson <5624115+spencerwilson@users.noreply.github.com>