feat(deps): Direct reqwest to use rustls for TLS (#124)
- Direct `reqwest` to use rustls w/ system certs for its TLS impl. This
removes the (currently undocumented) dependency on `openssl-sys` on
Linux platforms.

Branch created via `cargo add reqwest -F rustls-tls-native-roots`

The intended postcondition of this PR is that all TLS in the app is via
rustls and uses system certs. That was already the case for `tonic`, our
gRPC lib; this PR makes it true for our HTTP lib.

## Test plan

Worked offline w/ @bpiel , who verified that whereas on `main` the build
failed on openssl-sys, it succeeds on this branch.