fix(describe): Omit Snowflake's ARRAY, OBJECT, and VARIANT types (#154)
- Warn+omit ARRAY, OBJECT, and VARIANT-type columns during `dpm describe
snowflake`, as no codegen targets support them.
- Prefer explicit patterns instead of wildcard arm in codegen mappings
from universal type -> language type. This will make it harder to
unintentionally not cover a pattern in codegen.

PAT-4356 is a specialized version of a broader issue: There exist some
Snowflake types that can be `describe`'d but not code-gen'd, not only in
some targets but in _any_ targets. After auditing, it turns out the
problematic data types are the [semi-structured
types](https://docs.snowflake.com/en/sql-reference/data-types-semistructured):
- ARRAY
- OBJECT
- VARIANT

We definitely intend to support these better, but we don't currently and
the result when attempting to use them is poor: a panic during
`build-package`. This PR fixes that panic by replacing it with a
warn+omission from the descriptor during `describe`.

Further possible improvements on this code:
-
[PAT-3895](https://linear.app/patch-tech/issue/PAT-3895/finish-simplifying-the-descriptor):
Remove variants from
[`TableSchemaField`](https://github.com/patch-tech/dpm/blob/1bd20b948e34a0d1aa3fc7340798ebb81e0068e0/src/descriptor/table_schema.rs#L1221-L1598)
until the following is true: "If a column can be described, all targets
support it". In other words, there are no `unreachable!` branches in the
codegens' match expressions.